refactor(tsys_xml): keep models in transformer
